### 2022-06-22 #4498 @editorialbot
[PRE REVIEW]: GNU Data Language 1.0: a free/libre and open-source drop-in replacement for IDL/PV-WAVE
**Submitting author:** @pjb7687 (<a href="http://orcid.org/0000-0002-9064-4912">Jeongbin Park</a>)
**Repository:** https://github.com/gnudatalanguage/gdl
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
