### 2022-06-22 #4498 @Kevin-Mattheus-Moerman
@pjb7687 thanks for this submission to JOSS. I am the AEiC on call this week. I am here to help process this work for evaluation.

- [ ] Please check our [paper preparation instuctions](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), the above paper compilation error relations to the affiliations, please format them as shown in the example. You can call `@editorialbot generate pdf` here to update the paper.
- [ ] Please check those potentially missing DOI's :point_up:, you can add them to your bib file and rerun the paper compilation. Also you can call `@editorialbot check references` to check the DOIs again.
