### 2022-06-22 71c4d97
use GetWriteablePar() simplified poly
