+++ src/plotting_oplot.cpp
@@ -58,7 +58,7 @@ namespace lib {
 
       //test and transform eventually if POLAR and/or NSUM!
       if (nParam() == 1) {
-        yTemp = e->GetParAs< DDoubleGDL>(0);
+        yTemp = e->GetWriteableParAs< DDoubleGDL>(0);
         if (yTemp->Rank() == 0)
           e->Throw("Expression must be an array in this context: " + e->GetParString(0));
         yEl = yTemp->N_Elements();
@@ -66,11 +66,11 @@ namespace lib {
         xTemp = new DDoubleGDL(dimension(xEl), BaseGDL::INDGEN);
         xtemp_guard.Reset(xTemp); // delete upon exit
       } else {
-        xTemp = e->GetParAs< DDoubleGDL>(0);
+        xTemp = e->GetWriteableParAs< DDoubleGDL>(0);
         if (xTemp->Rank() == 0)
           e->Throw("Expression must be an array in this context: " + e->GetParString(0));
         xEl = xTemp->N_Elements();
-        yTemp = e->GetParAs< DDoubleGDL>(1);
+        yTemp = e->GetWriteableParAs< DDoubleGDL>(1);
         if (yTemp->Rank() == 0)
           e->Throw("Expression must be an array in this context: " + e->GetParString(1));
         yEl = yTemp->N_Elements();
@@ -217,7 +217,7 @@ namespace lib {
             for (auto i = 0; i < npts; ++i) (*x)[i] = (*lonlat)[i];
             DDoubleGDL* y = new DDoubleGDL(dimension(npts), BaseGDL::NOZERO);
             for (auto i = 0; i < npts; ++i) (*y)[i] = (*lonlat)[i + npts];
-            draw_polyline(actStream, x, y, 0, 0, false, false, false, psym, false, doColor ? color : NULL);
+            draw_polyline(actStream, x, y, psym, false, doColor ? color : NULL); //x and y are normed values.
             GDLDelete(x);
             GDLDelete(y);
           }
@@ -231,7 +231,7 @@ namespace lib {
         gdlGetAxisType(XAXIS, xLog);
         gdlGetAxisType(YAXIS, yLog);
         SelfConvertToNormXY(xVal, xLog, yVal, yLog, coordinateSystem); //DATA
-        draw_polyline(actStream, xVal, yVal, 0, 0, false, xLog, yLog, psym, false, doColor ? color : NULL);
+        draw_polyline(actStream, xVal, yVal, psym, false, doColor ? color : NULL); //x and y are normed values.
       }
     } //end of call_oplot
 
