### 2022-06-22 0d9cfc8
an attempt to fix JOSS paper compilation
