### 2022-06-22 c6ac6b0
fix more JOSS paper afiiliation issues
