### 2022-06-22 5dc7820
add missing DOIs
