+++ paper/paper.bib
@@ -96,7 +96,8 @@ archivePrefix = {arXiv},
   number={1},
   pages={1--13},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41598-021-92923-4}
 }
 
 @article{coates2019high,
@@ -107,7 +108,8 @@ archivePrefix = {arXiv},
   number={9},
   pages={eaax2742},
   year={2019},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/sciadv.aax2742}
 }
 
 @article{xiao2022superheating,
@@ -118,7 +120,8 @@ archivePrefix = {arXiv},
   number={1},
   pages={1--9},
   year={2022},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-022-29254-z}
 }
 
 @article{lee2022transboundary,
@@ -129,7 +132,8 @@ archivePrefix = {arXiv},
   number={1},
   pages={1--10},
   year={2022},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41598-022-08854-1}
 }
 
 @article{castro2021very,
@@ -140,7 +144,8 @@ archivePrefix = {arXiv},
   number={7890},
   pages={621--624},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1016/j.devcel.2021.10.006}
 }
 
 @article{chen2021orf3a,
