+++ paper/paper.bib
@@ -25,8 +25,7 @@
        editor = {{Mizumoto}, Y. and {Morita}, K. -I. and {Ohishi}, M.},
        series = {Astronomical Society of the Pacific Conference Series},
        volume = {434},
-archivePrefix = {arXiv},
-       eprint = {1101.0679}
+          doi = {10.48550/arXiv.1101.0679}
 }
 
 @book{Gumley_2010,
@@ -84,8 +83,7 @@ archivePrefix = {arXiv},
        editor = {{Ballester}, Pascal and {Ibsen}, Jorge and {Solar}, Mauricio and {Shortridge}, Keith},
        series = {Astronomical Society of the Pacific Conference Series},
        volume = {522},
-archivePrefix = {arXiv},
-       eprint = {1909.02371},
+          doi = {10.48550/arXiv.1909.02371}
 }
 
 @article{burrage2021cardiac,
@@ -145,7 +143,7 @@ archivePrefix = {arXiv},
   pages={621--624},
   year={2021},
   publisher={Nature Publishing Group},
-  doi={10.1016/j.devcel.2021.10.006}
+  doi={10.1038/s41586-021-04101-1},
 }
 
 @article{chen2021orf3a,
@@ -156,5 +154,6 @@ archivePrefix = {arXiv},
   number={23},
   pages={3250--3263},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.devcel.2021.10.006}
 }
