+++ src/plotting_axis.cpp
@@ -18,7 +18,6 @@
 #include "includefirst.hpp"
 #include "plotting.hpp"
 
-#define EXCH(a,b) { if (a>b){DDouble tmp=a; a=b; b=tmp;} a=MAX(a,0); a=MIN(a,1); b=MAX(b,0); b=MIN(b,1);}
 namespace lib {
 
   using namespace std;
@@ -26,13 +25,13 @@ namespace lib {
   class axis_call : public plotting_routine_call {
     DDouble xPos, yPos, zPos;
     bool doT3d;
-    bool standardNumPos;
     bool xAxis, yAxis, zAxis;
     COORDSYS coordinateSystem = DATA;
     // will handle axis logness..
     bool xLog, yLog, zLog;
     // is current box log or not?
     bool xAxisWasLog, yAxisWasLog, zAxisWasLog;
+    DLong xaxis_value, yaxis_value, zaxis_value;
     //in absence of arguments we will have:
     bool ynodef = true;
     bool xnodef = true;
@@ -55,38 +54,21 @@ namespace lib {
       static int yaxisIx = e->KeywordIx("YAXIS");
       static int zaxisIx = e->KeywordIx("ZAXIS");
 
-      PLINT xaxis_value, yaxis_value, zaxis_value;
       //IDL behaviour for XAXIS and YAXIS and ZAXIS options: only one option is considered, and ZAXIS above YAXIS above XAXIS
-      standardNumPos = true;
       if ((e->GetKW(xaxisIx) != NULL)) {
         xAxis = true;
         e->AssureLongScalarKWIfPresent(xaxisIx, xaxis_value);
-        if (xaxis_value == 0) {
-          standardNumPos = true;
-        } else {
-          standardNumPos = false;
-        }
       }
       if (e->GetKW(yaxisIx) != NULL) {
         yAxis = true;
         xAxis = false; // like in IDL, yaxis overrides xaxis
         e->AssureLongScalarKWIfPresent(yaxisIx, yaxis_value);
-        if (yaxis_value == 0) {
-          standardNumPos = true;
-        } else {
-          standardNumPos = false;
-        }
       }
       if (e->GetKW(zaxisIx) != NULL) {
         zAxis = true;
         xAxis = false;
         yAxis = false; // like in IDL, zaxis overrides all
         e->AssureLongScalarKWIfPresent(zaxisIx, zaxis_value);
-        if (zaxis_value == 0) {
-          standardNumPos = true;
-        } else {
-          standardNumPos = false;
-        }
       }
 
       static int DATAIx = e->KeywordIx("DATA");
@@ -120,15 +102,20 @@ namespace lib {
       gdlGetAxisType(ZAXIS, zAxisWasLog);
       
       GetCurrentUserLimits(oxStart,oxEnd, oyStart, oyEnd, ozStart, ozEnd);
+      if (ozStart==0 && ozEnd-ozStart==0) ozEnd=1; //this may happen
+      if (xAxisWasLog) {oxStart=pow(10,oxStart);oxEnd=pow(10,oxEnd);}
+      if (yAxisWasLog) {oyStart=pow(10,oyStart);oyEnd=pow(10,oyEnd);}
+      if (zAxisWasLog) {ozStart=pow(10,ozStart);ozEnd=pow(10,ozEnd);}
+      
       gdlAdjustAxisRange(e, XAXIS, oxStart, oxEnd, xAxisWasLog); //Box adjustement
       gdlAdjustAxisRange(e, YAXIS, oyStart, oyEnd, yAxisWasLog); //Box adjustement
       gdlAdjustAxisRange(e, ZAXIS, ozStart, ozEnd, zAxisWasLog); //Box adjustement
 
[…]
