### 2022-06-24 a554026
adding box_cursor and color_convert (doing nothing)
