### 2022-06-25 #4498 @Kevin-Mattheus-Moerman
@slayoo thanks for fixing those DOIs, you still however need to the affiliations, the are required to be like this (i.e. numbered):

```
authors:
  - name: Adrian M. Price-Whelan
    orcid: 0000-0000-0000-0000
    equal-contrib: true
    affiliation: "1, 2" # (Multiple affiliations must be quoted)
  - name: Author Without ORCID
    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
    affiliation: 2
  - name: Author with no affiliation
    corresponding: true # (This is how to denote the corresponding author)
    affiliation: 3
affiliations:
 - name: Lyman Spitzer, Jr. Fellow, Princeton University, USA
   index: 1
 - name: Institution Name, Country
   index: 2
 - name: Independent Researcher, Country
   index: 3
```
See: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
