+++ src/plotting_shade_surf.cpp
@@ -19,6 +19,8 @@
 #include "plotting.hpp"
 #include "dinterpreter.hpp"
 
+static GDL_3DTRANSFORMDEVICE T3DForAXes;
+
 namespace lib
 {
   using namespace std;
@@ -31,22 +33,37 @@ namespace lib
     DDoubleGDL *zVal, *yVal, *xVal;
     Guard<BaseGDL> xval_guard, yval_guard, zval_guard, p0_guard;
     SizeT xEl, yEl, zEl;
-    DDouble xStart, xEnd, yStart, yEnd, zStart, zEnd, datamax, datamin;
+    DDouble xStart, xEnd, yStart, yEnd, zStart, zEnd, zValue, datamax, datamin;
+    DDouble minVal,maxVal;
+    bool hasMinVal, hasMaxVal;
+    bool doT3d;
     bool nodata;
     bool setZrange;
     bool xLog;
     bool yLog;
     bool zLog;
     T3DEXCHANGECODE axisExchangeCode;
-    DLongGDL *shades;
-    Guard<BaseGDL> shades_guard;
-    bool doShade;
+    PLFLT alt = 30.0;
+    PLFLT az = 30.0;
+    PLFLT ay = 0;
+    
  private:
     bool handle_args (EnvT* e)
     {
+      //T3D ?
+      static int t3dIx = e->KeywordIx("T3D");
+      doT3d = (e->BooleanKeywordSet(t3dIx) || T3Denabled());
+
+      zValue = std::numeric_limits<DDouble>::quiet_NaN(); //NAN = no zValue?
+      static int zvIx = e->KeywordIx("ZVALUE");
+      if (e->KeywordPresent(zvIx)){
+        e->AssureDoubleScalarKWIfPresent(zvIx, zValue);
+        zValue = min(zValue, ZVALUEMAX); //to avoid problems with plplot
+        zValue = max(zValue, 0.0);
+      }
+      
      // in all cases, we have to exit here
-      if ( nParam()==2 || nParam()>3 )
-	{
+      if ( nParam()==2 || nParam()>3 )	{
 	  e->Throw ( "Incorrect number of arguments." );
 	}
 
@@ -59,35 +76,19 @@ namespace lib
     static int xLogIx = e->KeywordIx( "XLOG" );
     static int yLogIx = e->KeywordIx( "YLOG" );
     static int zLogIx = e->KeywordIx( "ZLOG" );
-    static int xTickunitsIx = e->KeywordIx( "XTICKUNITS" );
-    static int yTickunitsIx = e->KeywordIx( "YTICKUNITS" );
-    static int zTickunitsIx = e->KeywordIx( "ZTICKUNITS" );
 
     if ( e->KeywordPresent( xTypeIx ) ) xLog = e->KeywordSet( xTypeIx ); else xLog = e->KeywordSet( xLogIx );
     if ( e->KeywordPresent( yTypeIx ) ) yLog = e->KeywordSet( yTypeIx ); else yLog = e->KeywordSet( yLogIx );
     if ( e->KeywordPresent( zTypeIx ) ) zLog = e->KeywordSet( zTypeIx ); else zLog = e->KeywordSet( zLogIx );
 
-    if ( xLog && e->KeywordSet( xTickunitsIx ) ) {
-      Message( "PLOT: LOG setting ignored for Date/Time TICKUNITS." );
-      xLog = FALSE;
-    }
-    if ( yLog && e->KeywordSet( yTickunitsIx ) ) {
-      Message( "PLOT: LOG setting ignored for Date/Time TICKUNITS." );
-      yLog = FALSE;
-    }
-    if ( zLog && e->KeywordSet( zTickunitsIx ) ) {
[…]
