+++ src/devicewx.hpp
@@ -182,7 +182,7 @@ long style = (wxMINIMIZE_BOX | wxMAXIMIZE_BOX | wxRESIZE_BORDER | wxCAPTION | wx
   }
   plotFrame->Refresh();
   plotFrame->Update();
-//  plotFrame->Raise();
+  plotFrame->Raise();
 
 //really show by letting the loop do its magic. Necessary.
 #ifdef __WXMAC__
