### 2022-06-30 23c0a24
a few ccpcheck suggested improvements
