### 2022-07-01 81a76cd
contour supports projections and has ok levels wrt idl
