+++ src/plotting_contour.cpp
@@ -21,7 +21,7 @@
 #include "math_utl.hpp"
 
 #define LABELOFFSET 0.003
-#define LABELSPACING 0.1
+#define LABELSPACING 0.4
 
 // sometime after 10/14 the tr0,tr1 moved from plstream:: to plcallback::
 
@@ -47,6 +47,7 @@ namespace lib {
     fprintf(stderr, "x=%f,y=%f\n", x, y);
   }
   //kept out of object vaguely because plplot is in C, but should not be the case.
+
   PLINT doIt(PLFLT x, PLFLT y) {
     if (xLog && x <= 0) return 0;
     if (yLog && y <= 0) return 0;
@@ -62,17 +63,17 @@ namespace lib {
     DDouble minVal, maxVal;
     bool isLog;
     bool overplot, nodata;
-    COORDSYS coordinateSystem = DATA;
     DLongGDL *colors, *labels, *style;
     DFloatGDL* thick;
     Guard<BaseGDL> colors_guard, thick_guard, labels_guard, style_guard;
     DFloatGDL *spacing, *orientation;
     Guard<BaseGDL> spacing_guard, orientation_guard;
-    bool doT3d, flat3d;
+    bool doT3d = false;
+    bool flat3d = false;
     bool irregular;
     bool iso;
     DDouble zValue;
-    bool hasMinVal,hasMaxVal;
+    bool hasMinVal, hasMaxVal;
     bool recordPath;
 
     //PATH_XY etc: use actStream->stransform with a crafted recording function per level [lev-maxmax].
@@ -82,19 +83,16 @@ namespace lib {
     bool handle_args(EnvT* e) {
       //T3D ?
       static int t3dIx = e->KeywordIx("T3D");
-      doT3d = (e->BooleanKeywordSet(t3dIx) || T3Denabled());
+      doT3d = (e->BooleanKeywordSet(t3dIx) || T3Denabled()); //should check !P.T is OK 
 
       zValue = std::numeric_limits<DDouble>::quiet_NaN(); //NAN = no zValue?
       static int zvIx = e->KeywordIx("ZVALUE");
-      if (e->KeywordPresent(zvIx)){
+      if (doT3d && e->KeywordPresent(zvIx)) {
         e->AssureDoubleScalarKWIfPresent(zvIx, zValue);
         zValue = min(zValue, ZVALUEMAX); //to avoid problems with plplot
         zValue = max(zValue, 0.0);
-        flat3d=true;
+        flat3d = true;
       }
-      //zStart and zEnd are ALWAYS Zero. zValue will be registered in !Z.REGION
-      zStart = 0;
-      zEnd = 0;
 
       static int irregIx = e->KeywordIx("IRREGULAR");
       irregular = e->KeywordSet(irregIx);
@@ -114,8 +112,7 @@ namespace lib {
       if (nParam() == 1) {
         if (irregular) {
           e->Throw("Incorrect number of arguments.");
-        }
-        else {
+        } else {
           BaseGDL* p0 = e->GetNumericArrayParDefined(0)->Transpose(NULL);
           p0_guard.Init(p0); // delete upon exit
 
@@ -133,16 +130,16 @@ namespace lib {
       }
 
       if (nParam() == 3) {
+       //ZVal will be treated as 1 dim array and X and Y must have the same number of elements.
+        BaseGDL* p0 = e->GetNumericArrayParDefined(0)->Transpose(NULL);
+        p0_guard.Init(p0); // delete upon exit
+
+        zVal = static_cast<DDoubleGDL*> (p0->Convert2(GDL_DOUBLE, BaseGDL::COPY));
+        zval_guard.Init(zVal); // delete upon exit
         if (irregular) {
[…]

[…]
