### 2022-07-02 9e46b2b
solved some plplot crashes (ugly patch!)
