### 2022-07-02 1bc8271
JOSS paper changes from Alain
