+++ paper/paper.md
@@ -7,96 +7,88 @@ tags:
   - IDL
   - PV-WAVE
 authors:
-  - name: Jeongbin Park
+  - name: Jeongbin&nbsp;Park
     orcid: 0000-0002-9064-4912
-    affiliation: 1
-  - name: Gilles Duvert
+    affiliation: pusan.ac.kr
+  - name: Gilles&nbsp;Duvert
     orcid: 0000-0001-8769-3660 
-    affiliation: 2
-  - name: Alain Coulais
+    affiliation: univ-grenoble-alpes.fr
+  - name: Alain&nbsp;Coulais
     orcid: 0000-0001-6492-7719
-    affiliation: "3, 4"
-  - name: Gregory V. Jung
-    affiliation: 5
-  - name: Sylwester Arabas
+    affiliation: "obspm.fr,cea.fr"
+  - name: Gregory&nbsp;V.&nbsp;Jung
+    affiliation: ".us"
+  - name: Sylwester&nbsp;Arabas
     orcid: 0000-0003-2361-0082
-    affiliation: "6, 7"
+    affiliation: "illinois.edu,uj.edu.pl"
   - name: Brian Barker
-    affiliation: 8
-  - name: Takeshi Enomoto
-    affiliation: 9
-  - name: Sylvain Flinois
-    affiliation: 10
+    affiliation: ursinus.edu
+  - name: Takeshi&nbsp;Enomoto
+    affiliation: kyoto-u.ac.jp
+  - name: Sylvain&nbsp;Flinois
+    affiliation: kumullus.com
   - name: Oliver Gressel
-    affiliation: 11
+    affiliation: aip.de
   - name: Tomas Hillberg
-    affiliation: 12
+    affiliation: astro.su.se
   - name: Thibault Huillet
-    affiliation: 13
+    affiliation: .fr
   - name: Jan Kohnert
-    affiliation: 14
-  - name: Orion Poplawski
-    affiliation: 15
-  - name: Eloi Rozier de Linage
-    affiliation: 3
-  - name: Remi A. Solås
-    affiliation: 16
-  - name: Luke Stagner 
+    affiliation: ".de"
+  - name: Orion&nbsp;Poplawski
+    affiliation: nwra.com
+  - name: Eloi&nbsp;Rozier&nbsp;de&nbsp;Linage
+    affiliation: obspm.fr
+  - name: Remi&nbsp;A.&nbsp;Solås
+    affiliation: ".no" 
+  - name: Luke&nbsp;Stagner 
     orcid: 0000-0001-5516-3729
-    affiliation: 17
-  - name: Ole Streicher
+    affiliation: ga.com
+  - name: Ole&nbsp;Streicher
     orcid: 0000-0001-7751-1843
-    affiliation: 11
-  - name: James Tappin
-    affiliation: 18
-  - name: Thierry Thomas
-    affiliation: 19
-  - name: Jingwei Wang
-    affiliation: 20
-  - name: Christian Wimmer
-    affiliation: 14
+    affiliation: aip.de
+  - name: James&nbsp;Tappin
+    affiliation: stfc.ac.uk
+  - name: Thierry&nbsp;Thomas
+    affiliation: edf.fr
+  - name: Jingwei&nbsp;Wang
+    affiliation: mines-paristech.fr
+  - name: Christian&nbsp;Wimmer
+    affiliation: ".de"
 
 affiliations:
- - name: School of Biomedical Convergence Engineering, Pusan National University, Republic of Korea
-   index: 1
+ - name: Pusan National University, Korea
+   index: pusan.ac.kr
  - name: Univ. Grenoble Alpes, CNRS, IPAG, 38000 Grenoble, France 
-   index: 2
[…]
