+++ paper/paper.bib
@@ -25,7 +25,8 @@
        editor = {{Mizumoto}, Y. and {Morita}, K. -I. and {Ohishi}, M.},
        series = {Astronomical Society of the Pacific Conference Series},
        volume = {434},
-          doi = {10.48550/arXiv.1101.0679}
+archivePrefix = {arXiv},
+       eprint = {1101.0679}
 }
 
 @book{Gumley_2010,
@@ -66,7 +67,7 @@
 }
 
 @inproceedings{Coulais_et_al_2019,
-       author = {{Coulais}, Alain and {Duvert}, G. and {Jung}, G. and {Arabas}, S. and {Flinois}, S. and {Lounis}, A.~S.},
+       author = {{Coulais}, A. and {Duvert}, G. and {Jung}, G. and {Arabas}, S. and {Flinois}, S. and {Lounis}, A.~S.},
         title = "{GDL - GNU Data Language 0.9.9}",
     booktitle = {Astronomical Data Analysis Software and Systems XXVII},
          year = 2019,
@@ -76,14 +77,15 @@
 }
 
 @inproceedings{Duvert_et_al_2020,
-       author = {{Duvert}, Gilles and {Coulais}, Alain and {Schellens}, Mark},
+       author = {{Duvert}, G. and {Coulais}, A. and {Schellens}, M.},
         title = "{GDL Today: Reaching a Viable Alternative to IDL}",
     booktitle = {Astronomical Data Analysis Software and Systems XXVII},
          year = 2020,
        editor = {{Ballester}, Pascal and {Ibsen}, Jorge and {Solar}, Mauricio and {Shortridge}, Keith},
        series = {Astronomical Society of the Pacific Conference Series},
        volume = {522},
-          doi = {10.48550/arXiv.1909.02371}
+archivePrefix = {arXiv},
+       eprint = {1909.02371},
 }
 
 @article{burrage2021cardiac,
@@ -94,8 +96,7 @@
   number={1},
   pages={1--13},
   year={2021},
-  publisher={Nature Publishing Group},
-  doi={10.1038/s41598-021-92923-4}
+  publisher={Nature Publishing Group}
 }
 
 @article{coates2019high,
@@ -106,8 +107,7 @@
   number={9},
   pages={eaax2742},
   year={2019},
-  publisher={American Association for the Advancement of Science},
-  doi={10.1126/sciadv.aax2742}
+  publisher={American Association for the Advancement of Science}
 }
 
 @article{xiao2022superheating,
@@ -118,8 +118,7 @@
   number={1},
   pages={1--9},
   year={2022},
-  publisher={Nature Publishing Group},
-  doi={10.1038/s41467-022-29254-z}
+  publisher={Nature Publishing Group}
 }
 
 @article{lee2022transboundary,
@@ -130,8 +129,7 @@
   number={1},
   pages={1--10},
   year={2022},
-  publisher={Nature Publishing Group},
-  doi={10.1038/s41598-022-08854-1}
+  publisher={Nature Publishing Group}
 }
 
 @article{castro2021very,
@@ -142,8 +140,7 @@
   number={7890},
   pages={621--624},
   year={2021},
-  publisher={Nature Publishing Group},
-  doi={10.1038/s41586-021-04101-1},
+  publisher={Nature Publishing Group}
 }
 
 @article{chen2021orf3a,
@@ -154,6 +151,51 @@
   number={23},
   pages={3250--3263},
   year={2021},
-  publisher={Elsevier},
-  doi={10.1016/j.devcel.2021.10.006}
+  publisher={Elsevier}
 }
+
+@ARTICLE{Danielski2018AJ,
[…]
