### 2022-07-04 32bfd69
fix affiliations and bibliographies
