+++ paper/paper.bib
@@ -25,8 +25,7 @@
        editor = {{Mizumoto}, Y. and {Morita}, K. -I. and {Ohishi}, M.},
        series = {Astronomical Society of the Pacific Conference Series},
        volume = {434},
-archivePrefix = {arXiv},
-       eprint = {1101.0679}
+          doi = {10.48550/arXiv.1101.0679}
 }
 
 @book{Gumley_2010,
@@ -67,7 +66,7 @@ archivePrefix = {arXiv},
 }
 
 @inproceedings{Coulais_et_al_2019,
-       author = {{Coulais}, A. and {Duvert}, G. and {Jung}, G. and {Arabas}, S. and {Flinois}, S. and {Lounis}, A.~S.},
+       author = {{Coulais}, Alain and {Duvert}, G. and {Jung}, G. and {Arabas}, S. and {Flinois}, S. and {Lounis}, A.~S.},
         title = "{GDL - GNU Data Language 0.9.9}",
     booktitle = {Astronomical Data Analysis Software and Systems XXVII},
          year = 2019,
@@ -77,15 +76,14 @@ archivePrefix = {arXiv},
 }
 
 @inproceedings{Duvert_et_al_2020,
-       author = {{Duvert}, G. and {Coulais}, A. and {Schellens}, M.},
+       author = {{Duvert}, Gilles and {Coulais}, Alain and {Schellens}, Mark},
         title = "{GDL Today: Reaching a Viable Alternative to IDL}",
     booktitle = {Astronomical Data Analysis Software and Systems XXVII},
          year = 2020,
        editor = {{Ballester}, Pascal and {Ibsen}, Jorge and {Solar}, Mauricio and {Shortridge}, Keith},
        series = {Astronomical Society of the Pacific Conference Series},
        volume = {522},
-archivePrefix = {arXiv},
-       eprint = {1909.02371},
+          doi = {10.48550/arXiv.1909.02371}
 }
 
 @article{burrage2021cardiac,
@@ -96,7 +94,8 @@ archivePrefix = {arXiv},
   number={1},
   pages={1--13},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41598-021-92923-4}
 }
 
 @article{coates2019high,
@@ -107,7 +106,8 @@ archivePrefix = {arXiv},
   number={9},
   pages={eaax2742},
   year={2019},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/sciadv.aax2742}
 }
 
 @article{xiao2022superheating,
@@ -118,7 +118,8 @@ archivePrefix = {arXiv},
   number={1},
   pages={1--9},
   year={2022},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-022-29254-z}
 }
 
 @article{lee2022transboundary,
@@ -129,7 +130,8 @@ archivePrefix = {arXiv},
   number={1},
   pages={1--10},
   year={2022},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41598-022-08854-1}
 }
 
 @article{castro2021very,
@@ -140,7 +142,8 @@ archivePrefix = {arXiv},
   number={7890},
   pages={621--624},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41586-021-04101-1},
 }
 
 @article{chen2021orf3a,
[…]
