+++ src/plotting.hpp
@@ -246,7 +246,48 @@ namespace lib {
   void gdlSetZto3DDriverTransform( GDLGStream *a, DDouble zValue);
   void gdlStop3DDriverTransform(GDLGStream *a);
   bool T3Denabled();
-
+  void gdlDoRangeExtrema(DDoubleGDL *xVal, DDoubleGDL *yVal, DDouble &min, DDouble &max, DDouble xmin, DDouble xmax, bool doMinMax = false, DDouble minVal = 0, DDouble maxVal = 0);
+  void draw_polyline(GDLGStream *a, DDoubleGDL *xVal, DDoubleGDL *yVal, DLong psym = 0, bool append = false, DLongGDL *color = NULL);
+  void SelfNormLonLat(DDoubleGDL *lonlat);
+  void SelfPDotTTransformProjectedPolygonTable(DDoubleGDL *lonlat);
+  DDoubleGDL* GDLgrGetProjectPolygon(GDLGStream * a, PROJTYPE ref, DStructGDL* map, DDoubleGDL *lons, DDoubleGDL *lats, DDoubleGDL *zVal, bool isRadians, bool const doFill, bool const dolines, DLongGDL *&conn);
+  void GDLgrPlotProjectedPolygon(GDLGStream * a, DDoubleGDL *lonlat, bool const doFill, DLongGDL *conn);
+  void gdlSetGraphicsPenColorToBackground(GDLGStream *a);
+  void gdlLineStyle(GDLGStream *a, DLong style);
+  DFloat* gdlGetRegion();
+  void gdlStoreXAxisParameters(GDLGStream* actStream, DDouble Start, DDouble End, bool log);
+  void gdlStoreYAxisParameters(GDLGStream* actStream, DDouble Start, DDouble End, bool log);
+  void gdlStoreZAxisParameters(GDLGStream* actStream, DDouble Start, DDouble End, bool log, DDouble zposStart, DDouble zposEnd);
+  void gdlGetAxisType(int axisId, bool &log);
+  void gdlGetCurrentAxisWindow(int axisId, DDouble &wStart, DDouble &wEnd);
+  void gdlStoreAxisType(int axisId, bool type);
+  //  void gdlGetCharSizes(GDLGStream *a, PLFLT &nsx, PLFLT &nsy, DDouble &wsx, DDouble &wsy, 
+  //		       DDouble &dsx, DDouble &dsy, DDouble &lsx, DDouble &lsy); 
+  void GetSFromPlotStructs(DDouble **sx, DDouble **sy, DDouble **sz = NULL);
+  void GetWFromPlotStructs(DDouble *wx, DDouble *wy, DDouble *wz = NULL);
+  void ConvertToNormXY(SizeT n, DDouble *x, bool const xLog, DDouble *y, bool const yLog, COORDSYS const code);
+  void ConvertToNormZ(SizeT n, DDouble *z, bool const zLog, COORDSYS const code);
+  void gdlStoreCLIP();
+//  void gdlGetCLIPXY(DLong &x0, DLong &y0, DLong &x1, DLong &y1);
+  void GetCurrentUserLimits(DDouble &xStart, DDouble &xEnd, DDouble &yStart, DDouble &yEnd); //2D
+  void GetCurrentUserLimits(DDouble &xStart, DDouble &xEnd, DDouble &yStart, DDouble &yEnd, DDouble &zStart, DDouble &zEnd); //3D
+  void gdlAdjustAxisRange(EnvT* e, int axisId, DDouble &val_min, DDouble &val_max, bool &log);
+  PLFLT AutoTick(DDouble x);
+  PLFLT AutoLogTick(DDouble min, DDouble max);
[…]
