+++ src/plotting.cpp
@@ -494,18 +494,27 @@ namespace lib
     if (sz != NULL) *sz= &(*static_cast<DDoubleGDL*>(zStruct->GetTag(szTag, 0)))[0];
   }
   
-//  void GetWFromPlotStructs(DFloat **wx, DFloat **wy, DFloat **wz )
-//  {
-//    DStructGDL* xStruct=SysVar::X();   //MUST NOT BE STATIC, due to .reset 
-//    DStructGDL* yStruct=SysVar::Y();   //MUST NOT BE STATIC, due to .reset 
-//    DStructGDL* zStruct=SysVar::Z();   //MUST NOT BE STATIC, due to .reset 
-//    unsigned xwindowTag=xStruct->Desc()->TagIndex("WINDOW");
-//    unsigned ywindowTag=yStruct->Desc()->TagIndex("WINDOW");
-//    unsigned zwindowTag=zStruct->Desc()->TagIndex("WINDOW");
-//    *wx= &(*static_cast<DFloatGDL*>(xStruct->GetTag(xwindowTag, 0)))[0];
-//    *wy= &(*static_cast<DFloatGDL*>(yStruct->GetTag(ywindowTag, 0)))[0];
-//    *wz= &(*static_cast<DFloatGDL*>(zStruct->GetTag(zwindowTag, 0)))[0];
-//  }
+  void GetWFromPlotStructs(DDouble *wx, DDouble *wy, DDouble *wz )
+  {
+    DStructGDL* xStruct=SysVar::X();   //MUST NOT BE STATIC, due to .reset 
+    DStructGDL* yStruct=SysVar::Y();   //MUST NOT BE STATIC, due to .reset 
+    DStructGDL* zStruct=SysVar::Z();   //MUST NOT BE STATIC, due to .reset 
+    unsigned xwindowTag=xStruct->Desc()->TagIndex("WINDOW");
+    unsigned ywindowTag=yStruct->Desc()->TagIndex("WINDOW");
+    unsigned zwindowTag=zStruct->Desc()->TagIndex("WINDOW");
+    if (wx != NULL) {
+      wx[0]= (*static_cast<DFloatGDL*>(xStruct->GetTag(xwindowTag, 0)))[0];
+      wx[1]= (*static_cast<DFloatGDL*>(xStruct->GetTag(xwindowTag, 0)))[1];
+    }
+    if (wy != NULL) {
+      wy[0]= (*static_cast<DFloatGDL*>(yStruct->GetTag(ywindowTag, 0)))[0];
+      wy[1]= (*static_cast<DFloatGDL*>(yStruct->GetTag(ywindowTag, 0)))[1];
+    }
+    if (wz != NULL) {
+      wz[0]= (*static_cast<DFloatGDL*>(zStruct->GetTag(zwindowTag, 0)))[0];
+      wz[1]= (*static_cast<DFloatGDL*>(zStruct->GetTag(zwindowTag, 0)))[1];
+    }
+  }
 
   void GetUsym(DLong **n, DInt **do_fill, DFloat **x, DFloat **y, bool **do_color, DLong **usymColor , bool **do_thick, DFloat **usymThick)
   {
@@ -1145,15 +1154,11 @@ namespace lib
     //easier to retrieve here the values sent to vpor() in the calling function instead of
     //calling a special function like gdlStoreRegion(), gdlStoreWindow() each time a memory of vpor() is needed.
     //Will thus need to be amended when/if we get rid of plplot.
-    PLFLT p_xmin, p_xmax, p_ymin, p_ymax, norm_min, norm_max, charDim;
+    PLFLT p_xmin, p_xmax, p_ymin, p_ymax, norm_min, norm_max;
[…]

[…]
