### 2022-07-07 926f822
insure 'local' definition of MIN mAX and ABS; corrected some compiler warnings.
