### 2022-07-09 ac7d486
enable with a trick , surface for ax < 0 (impossible with plplot!) limits of the trick attained with shade_surf however.
