### 2022-07-10 5aa77f9
permit wxUSE_GRAPHICS_CONTEXT since it permits at least to erase parts of plts by rewriting them with !P.BACKGROUND
