### 2022-07-10 f93ee07
Create simple modal base, using wxWindowDisabler(). Probably too simple, but enables the use of cw_form() when not attached to an already existing base.
