### 2022-07-10 d7d7605
removing debug mode
