### 2022-07-11 bbd4e6f
Remove warning now that modal KW works.
