### 2022-07-12 1453165
remove regression: CLIP is OK probably speededed things up with test on polygon intersecting the ll_box and uv_box
