+++ src/projections.cpp
@@ -993,6 +993,10 @@ namespace lib {
     int inside; // number of polygons inside
     int outside; // number of polygons outside
     bool valid; //to be ignored (polygon has been transferred to another polygon list
+    DDouble xmin;
+    DDouble ymin;
+    DDouble xmax;
+    DDouble ymax;
   };
 
   struct Point3d {
@@ -1607,7 +1611,27 @@ namespace lib {
       pz = (*pipeline)[pipedims[0] * line + 7];
     }
   }
-
+  inline bool intersectsLonLatBox(DDouble minlon, DDouble maxlon, DDouble minlat, DDouble maxlat, DDouble* llbox){
+    if (minlon >= llbox[3]) return false;
+    if (maxlon <= llbox[1]) return false;
+    if (minlat >= llbox[2]) return false;
+    if (maxlat <= llbox[0]) return false;
+    return true;
+  }
+  inline bool intersectsBox(DDouble xmin, DDouble xmax, DDouble ymin, DDouble ymax, DDouble* box){
+    if (xmin >= box[3]) return false;
+    if (xmax <= box[1]) return false;
+    if (ymin >= box[2]) return false;
+    if (ymax <= box[0]) return false;
+    return true;
+  }
+  inline bool inBox(DDouble x, DDouble y, DDouble* box){
+    if (x > box[3]) return false;
+    if (x < box[1]) return false;
+    if (y > box[2]) return false;
+    if (y < box[0]) return false;
+    return true;
+  }
   //reproject, removing hidden points, cutting lines at projections splits & boundaries if LINE and FILL, closing contours if FILL.
   //produces a [3,N] or [N,3] output. Or -1 if nothing works.
   DDoubleGDL* gdlProjForward(PROJTYPE ref, DStructGDL* map, DDoubleGDL *lons, DDoubleGDL *lats, DDoubleGDL *zIn, DLongGDL *connIn,
@@ -1627,9 +1651,12 @@ namespace lib {
     unsigned llboxTag = map->Desc()->TagIndex("LL_BOX");
     DDoubleGDL* llbox = (static_cast<DDoubleGDL*> (map->GetTag(llboxTag, 0))->Dup());
     Guard<BaseGDL> llboxGuard(llbox);
+    unsigned uvboxTag = map->Desc()->TagIndex("UV_BOX");
+    DDoubleGDL* uvbox = (static_cast<DDoubleGDL*> (map->GetTag(uvboxTag, 0))->Dup());
+    Guard<BaseGDL> uvboxGuard(uvbox);
     //test if we can eliminate some polygons as they are probably (this is the use of ll_box) not going to be seen at the end.
     //this has problems as ll_box is very crude and false for some projetions (satellite)
-    //    bool llsubset=!((*llbox)[0] <= -90.0 && (*llbox)[2] >= 90.0 && (*llbox)[1] <= -180.0 && (*llbox)[3] >= 180.0);
+    bool llsubset=!((*llbox)[0] <= -90.0 && (*llbox)[2] >= 90.0 && (*llbox)[1] <= -180.0 && (*llbox)[3] >= 180.0);
 
     // convert to radians
     for (int i = 0; i < 4; ++i) (*llbox)[i] *= DEG_TO_RAD;
@@ -1732,22 +1759,19 @@ namespace lib {
             currentVertexList.push_back(curr);
           }
         }
-        //        if (llsubset) {
-        //          bool keep=intersectsLonLatBox(minlon,maxlon,minlat,maxlat,&((*llbox)[0]));
-        //          if (keep) 
-        //          {
-        //            currentPol.VertexList = currentVertexList;
-        //            currentPol.type = 1; //before cut
-        //            PolygonList.push_back(currentPol);
-        //          } else {
[…]
