### 2022-07-12 4cf6ad8
Update() for wxwidgets is not needed and was slowing regrettably the plotting.
