### 2022-07-12 eff2eaf
polyfill OK for projections in 3D removed pb with uninitialized variable in map_proj_init
