### 2022-07-13 f396fc3
GDL passes test_coyote. test_coyote has been added to testsuite/interactive_tests, with same logic as test_map (wait for input or use timer with tictoc=xxx) corrected a prblem of unknown macro. Strange why contour() fails with optimisation by compiler. To be cured.
