+++ src/plotting.cpp
@@ -95,7 +95,7 @@ namespace lib
        //min and max of y if not NaN and in range [minVal, maxVal] if doMinMax=yes (min_value, max_value keywords)
        valy=(*yVal)[i];
        if ((doMinMax && (valy<minVal || valy>maxVal )) || isnan(valy)) continue;
-       else {if(k==0) {min=valy; max=valy;} else {min=gdlPlot_Min(min,valy); max=gdlPlot_Max(max,valy);}}
+       else {if(k==0) {min=valy; max=valy;} else {min=MIN(min,valy); max=MAX(max,valy);}}
        }
        k++;
     }
