### 2022-07-13 429e204
temporary solve problem of effect of wrong dpi on PS and mem. Improved a bit PostScript. cgFixPs cannot work on such PS files.
