### 2022-07-28 #4498 @mgalloy
@gkthiruvathukal I am a member of the [GDL GitHub organization](https://github.com/orgs/gnudatalanguage/people?query=galloy). I am not an active member and have not participated in this paper. Is that a conflict?
