### 2022-07-29 #4498 @slayoo
@gkthiruvathukal just to clarify, we are adding anyone involved in discussions on GDL on GitHub to the [ gnudatalanguage/gdlteam ](https://github.com/orgs/gnudatalanguage/teams/gdlteam) user group on GitHub, hence everyone from this user community becomes a member of the GDL organization on GitHub (63 members as of now, group created to be able to send notification to users).
