### 2022-07-30 #4498 @gkthiruvathukal
@mgalloy At first blush, this does not seem like a conflict to me. The real question is whether there would be a perceived conflict of interest. I am not seeing anything here that suggests there would be. What we are seeking is reviewers who can evaluate the submission and help the authors to improve it.  I think I am going to go ahead and assign you as a reviewer unless you think you cannot conduct a thorough and fair evaluation of the software.
