### 2022-08-01 #4633 @mohawk2
@gkthiruvathukal @pjb7687 I have completed my review as reflected by all the boxes being checked above, and believe the paper and software are good to go.

I did decide to finesse this point a bit:
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)

...in that the various papers citing usage of GDL (and the unit tests, etc) definitely meet this, so all is well - but I feel that adding a short snippet of real-world code in the generally excellent project README.md would be a good idea.
