+++ CMakeLists.txt
@@ -318,7 +318,11 @@ if(OPENMP)
       set(LIBRARIES ${LIBRARIES} ${OpenMP_CXX_FLAGS})
     endif()
     if(APPLE)
-      link_directories(/usr/local/lib)
+      if(${MACHINE_ARCH} STREQUAL "arm64")
+        link_directories(/opt/homebrew/lib)
+      else()
+        link_directories(/usr/local/lib)
+      endif()
     endif()
   else(OPENMP_FOUND)
     message(FATAL_ERROR "Your compiler does not support OpenMP or OpenMP was not found"
