### 2022-08-04 022d7ce
Apple silicon (concise codes)
