+++ scripts/build_gdl.sh
@@ -426,8 +426,8 @@ function configure_gdl {
         if [[ ${Platform} == "arm64" ]]; then
             export LIBRARY_PATH=$LIBRARY_PATH:/opt/homebrew/opt/llvm/lib
             CMAKE_ADDITIONAL_ARGS=( "-DREADLINEDIR=/opt/homebrew/opt/readline"
-+                                   "-DCMAKE_CXX_COMPILER=/opt/homebrew/opt/llvm/bin/clang++"
-+                                   "-DCMAKE_C_COMPILER=/opt/homebrew/opt/llvm/bin/clang" )
+                                    "-DCMAKE_CXX_COMPILER=/opt/homebrew/opt/llvm/bin/clang++"
+                                    "-DCMAKE_C_COMPILER=/opt/homebrew/opt/llvm/bin/clang" )
         else
             export LIBRARY_PATH=$LIBRARY_PATH:/usr/local/opt/llvm/lib
             CMAKE_ADDITIONAL_ARGS=( "-DREADLINEDIR=/usr/local/opt/readline"
