### 2022-08-04 5a45ff6
Apparently these changes make 1) the wxWidgest plots fast and 2) avoids the eventloop to be stuck and no plots appear. works on Windows, OSX and linux
