+++ src/gdlwxstream.cpp
@@ -127,15 +127,15 @@ void GDLWXStream::DoUpdate()
   if( this->valid && container != NULL) {
     container->RepaintGraphics();
     //will be updated by eventloop.
-//#ifdef __WXMAC__
-//  wxTheApp->Yield();
-//#else
-//  wxGetApp().MainLoop(); //central loop for wxEvents!
-//#endif
+#ifdef __WXMAC__
+  wxTheApp->Yield();
+#else
+  wxGetApp().MainLoop(); //central loop for wxEvents!
+#endif
   }
 }
 //We should not need "Update"
-void GDLWXStream::Update(){}
+void GDLWXStream::Update(){DoUpdate();}
 ////should be used when one does not recreate a wxstream each time size changes...
 void GDLWXStream::SetSize( wxSize s )
 {
