### 2022-08-05 7cdccf1
reverted parts related to improving the refreshing of plots. We must have independent threads and eventloops for all windows.
