+++ src/devicewx.hpp
@@ -120,7 +120,7 @@ public:
 #ifdef __WXMAC__
 long style = (wxMINIMIZE_BOX | wxMAXIMIZE_BOX | wxRESIZE_BORDER | wxCAPTION | wxCLOSE_BOX);
 #else
-long style = (wxMINIMIZE_BOX | wxMAXIMIZE_BOX | wxRESIZE_BORDER | wxCAPTION | wxCLOSE_BOX);// | wxFRAME_TOOL_WINDOW); //no focus 
+long style = (wxMINIMIZE_BOX | wxMAXIMIZE_BOX | wxRESIZE_BORDER | wxCAPTION | wxCLOSE_BOX | wxFRAME_TOOL_WINDOW); //no focus 
 #endif
   gdlwxPlotFrame* plotFrame = new gdlwxPlotFrame(titleWxString, wxPoint(xoff,yoff), wxDefaultSize, style, scrolled);
   // Associate a sizer immediately
@@ -170,28 +170,14 @@ long style = (wxMINIMIZE_BOX | wxMAXIMIZE_BOX | wxRESIZE_BORDER | wxCAPTION | wx
   plotFrame->Connect(wxEVT_CLOSE_WINDOW, wxCloseEventHandler(gdlwxPlotFrame::OnUnhandledClosePlotFrame));
 //  plotFrame->Connect(wxEVT_SIZE, wxSizeEventHandler(gdlwxPlotFrame::OnPlotSizeWithTimer));
   plotFrame->Connect(wxEVT_SIZE, wxSizeEventHandler(gdlwxPlotFrame::OnPlotWindowSize));
-
-  //do not do anything else here, since it may affect hidden windows behaviour
-
-  if (hide) {
-   winList[ wIx]->UnMapWindowAndSetPixmapProperty(); //needed: will set the "pixmap" property
-  } else {
-    //only for REALIZED windows, we can change some 'interactive' properties.
-  plotFrame->Realize();
-  plotFrame->Raise();
-//on wxMAC, frame will not appear if style is not exactly this (!!!)
-#ifdef __WXMAC__
-    plotFrame->Show();
-#else
-//Depending on the WINDOW MANAGER, you may or not find the new plot behind other wondows on your display. At least this insures that
-// the keyboard focus is NOT given to the window, since you do not want to lose the focus in the command line terminal.
-    plotFrame->ShowWithoutActivating();
-#endif
-  }
-  plotFrame->Refresh();
-  plotFrame->Update();
-//plotFrame->Raise(); 
-
+    plotFrame->Realize();
+    if (hide) {
+      winList[ wIx]->UnMapWindowAndSetPixmapProperty(); //needed: will set the "pixmap" property
+    } else {
+      //    plotFrame->Raise(); //Raise will ALWAYS put focus on plot window, that we do not want
+      plotFrame->Show(); //WithoutActivating();
+    }
+    plotFrame->UpdateWindowUI();
 //really show by letting the loop do its magic. Necessary.
 #ifdef __WXMAC__
   wxTheApp->Yield();
