### 2022-08-05 f1de757
permit (again?) [x|y|z]sty=8  (alain's message)
