+++ src/plotting.hpp
@@ -1536,20 +1536,23 @@ namespace lib {
   // as it is temporarily superseded by setting a new a->win().
   //this makes GDLAXIS independent of WIN, and help the whole code to be dependent only on VPOR which is the sole useful plplot command to really use.
   //ZAXIS will always be an YAXIS plotted with a special YZEXCH T3D matrix. So no special handling of ZAXIS here.
+
   static void gdlAxis(EnvT *e, GDLGStream *a, int axisId, DDouble Start, DDouble End, bool Log, DLong modifierCode = 0) {
-    if (Start==End) return;
-    if (Log && (Start<=0 ||End <=0)) return; //important protection 
-    DLong Style;
-    gdlGetDesiredAxisStyle(e, axisId, Style);
-    if ((Style & 4) == 4) return; //if we do not write the axis...
-    
+    if (Start == End) return;
+    if (Log && (Start <= 0 || End <= 0)) return; //important protection 
+    DLong AxisStyle;
+    gdlGetDesiredAxisStyle(e, axisId, AxisStyle);
+    if ((AxisStyle & 4) == 4) return; //if we do not write the axis...
+
     // we WILL plot something, so set temporarlily WIN accordingly
-    PLFLT owxmin,owxmax,owymin,owymax;
-    a->getCurrentWorldBox(owxmin,owxmax,owymin,owymax);
+    PLFLT owxmin, owxmax, owymin, owymax;
+    a->getCurrentWorldBox(owxmin, owxmax, owymin, owymax);
     if (axisId == XAXIS) {
-      if (Log) a->wind(log10(Start),log10(End),owymin,owymax); else a->wind(Start,End,owymin,owymax);
+      if (Log) a->wind(log10(Start), log10(End), owymin, owymax);
+      else a->wind(Start, End, owymin, owymax);
     } else {
-      if (Log) a->wind(owxmin,owxmax,log10(Start),log10(End)); else a->wind(owxmin,owxmax,Start,End);
+      if (Log) a->wind(owxmin, owxmax, log10(Start), log10(End));
+      else a->wind(owxmin, owxmax, Start, End);
     }
     static GDL_TICKDATA tdata;
     tdata.a = a;
@@ -1576,7 +1579,8 @@ namespace lib {
 
     //special values
     PLFLT OtherAxisSizeInMm;
-    if (axisId == XAXIS) OtherAxisSizeInMm = a->mmyPageSize()*(a->boxnYSize()); else  OtherAxisSizeInMm = a->mmxPageSize()*(a->boxnXSize());
+    if (axisId == XAXIS) OtherAxisSizeInMm = a->mmyPageSize()*(a->boxnYSize());
+    else OtherAxisSizeInMm = a->mmxPageSize()*(a->boxnXSize());
     DLong GridStyle;
     gdlGetDesiredAxisGridStyle(e, axisId, GridStyle);
     DLong Minor;
@@ -1613,7 +1617,7 @@ namespace lib {
     //ticklen in a percentage of box x or y size, to be expressed in mm 
     if (axisId == XAXIS) ticklen_in_mm = a->mmyPageSize()*(a->boxnYSize()) * ticklen_in_mm;
     else ticklen_in_mm = a->mmxPageSize()*(a->boxnXSize()) * ticklen_in_mm;
-    if (ticklen_in_mm > 100.) ticklen_in_mm=0; //PATCH to avoid PS and MEM device problem. Check why gspa() returns silly values. TBC 
+    if (ticklen_in_mm > 100.) ticklen_in_mm = 0; //PATCH to avoid PS and MEM device problem. Check why gspa() returns silly values. TBC 
[…]
