### 2022-08-07 46ef2a5
this version handles better the creation of our wxApp. A small step towards the running of the wxApp in a separate thread.
