### 2022-08-08 cc4fa65
win32 subtlety -- but this code is probably not useful anymore
