### 2022-08-08 59fc6d5
These changes make gdl work OK on windows with our new drivers and also with the proj package.
