### 2022-08-10 298c289
allow compilation with IM and GM. Solve  #1206
