+++ src/magick_cl.cpp
@@ -48,7 +48,6 @@
     if (notInitialized ) { \
       notInitialized = false; \
         Magick::InitializeMagick(NULL);  \
-        if ( QuantumDepth < 16) fprintf(stderr, "%% INFO: the %s library support only 8 bit images.\n", MagickPackageName);\
               }
 
 namespace lib {
@@ -78,7 +77,9 @@ namespace lib {
 
   Image *magick_image(EnvT *e, unsigned int mid) {
     if (gValid[mid] == 0) e->Throw("invalid ID.");
-    return gImage[mid];
+    Image* im=gImage[mid];
+//    std::cerr<<im->depth()<<std::endl;
+    return im;
   }
 
   unsigned int magick_image(EnvT* e, Image *imImage) {
@@ -382,7 +383,9 @@ namespace lib {
 #endif
       {
         DByteGDL *bImage = new DByteGDL(dim, BaseGDL::ZERO);
-        PixelPacket *pixel_cache = image->getPixels(0,0,columns,rows); //magick command, without it writePixels do NOTHING!
+#ifndef HAS_IMAGEMAGICK
+        MagickCore::PixelPacket *pixel_cache = image->getPixels(0,0,columns,rows); //magick command, without it writePixels do NOTHING!
+#endif
         image->writePixels(IndexQuantum,(unsigned char*)(bImage->DataAddr()));
         return bImage;
       }
@@ -393,7 +396,9 @@ namespace lib {
 #endif
       {
         DUIntGDL *bImage = new DUIntGDL(dim, BaseGDL::NOZERO);
-        PixelPacket *pixel_cache = image->getPixels(0,0,columns,rows); //magick command, without it writePixels do NOTHING!
+#ifndef HAS_IMAGEMAGICK
+        MagickCore::PixelPacket *pixel_cache = image->getPixels(0,0,columns,rows); //magick command, without it writePixels do NOTHING!
+#endif
         image->writePixels(IndexQuantum,(unsigned char*)(bImage->DataAddr())); //probably not good, impossible to test yet.
         return bImage;
       }
@@ -608,14 +613,18 @@ namespace lib {
             if (image->matte()) map = map + "A";
           }
         }
-        if (image->depth() == 8) {
+        if (image->depth() == 8 || GDLimage->Type()==GDL_BYTE || !(IntType(GDLimage->Type()))) { //only INTeger values > BYTE can be saved as 16 bits see WRITE_PNG help
           DByteGDL * bImage = static_cast<DByteGDL*> (GDLimage->Convert2(GDL_BYTE, BaseGDL::COPY));
           Guard<DByteGDL> bImageGuard(bImage);
+        // Ensure that there are no other references to this image.
+          image->modifyImage();
           image->read(columns, rows, map, CharPixel, &(*bImage)[0]);
-        } else if (image->depth() == 16) {
-          DUIntGDL * iImage = static_cast<DUIntGDL*> (GDLimage->Convert2(GDL_UINT, BaseGDL::COPY));
-          Guard<DUIntGDL> iImageGuard(iImage);
-          image->read(columns, rows, map, ShortPixel, &(*iImage)[0]);
+        } else if (image->depth() == 16) { //the only one able to support writng INTeger values > BYTES
+            DUIntGDL * iImage = static_cast<DUIntGDL*> (GDLimage->Convert2(GDL_UINT, BaseGDL::COPY));
+            Guard<DUIntGDL> iImageGuard(iImage);
+        // Ensure that there are no other references to this image.
+            image->modifyImage();
+            image->read(columns, rows, map, ShortPixel, &(*iImage)[0]);
[…]
