+++ CMakeLists.txt
@@ -601,11 +601,14 @@ if(MAGICK)
   mark_as_advanced(ImageMagick_EXECUTABLE_DIR ImageMagick_Magick++_INCLUDE_DIR ImageMagick_Magick++_LIBRARY
     ImageMagick_MagickCore_INCLUDE_DIR ImageMagick_MagickCore_LIBRARY ImageMagick_MagickWand_INCLUDE_DIR ImageMagick_MagickWand_LIBRARY)
   set(USE_MAGICK ${ImageMagick_FOUND})
+  set(HAS_IMAGEMAGICK ${ImageMagick_FOUND})
   if(ImageMagick_FOUND)
     find_program(MAGICKXXCONFIG Magick++-config)
     if(MAGICKXXCONFIG)
       execute_process(COMMAND ${MAGICKXXCONFIG} "--libs" OUTPUT_VARIABLE MAGICKXXCONFIGLIBS OUTPUT_STRIP_TRAILING_WHITESPACE)
-      set(LIBRARIES ${LIBRARIES} ${MAGICKXXCONFIGLIBS})     
+      set(LIBRARIES ${LIBRARIES} ${MAGICKXXCONFIGLIBS})
+      execute_process(COMMAND ${MAGICKXXCONFIG} "--cxxflags" OUTPUT_VARIABLE MAGICKXXCONFIGCXXFLAGS OUTPUT_STRIP_TRAILING_WHITESPACE)
+      set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${MAGICKXXCONFIGCXXFLAGS}")  
     else(MAGICKXXCONFIG)
       message(FATAL_ERROR "ImageMagick is required but was not found (Magick++-config).\n"
     "Use -DMAGICKDIR=DIR to specify the ImageMagick directory.\n"
@@ -614,7 +617,7 @@ if(MAGICK)
         "(suitable Fedora/CentOS package: ImageMagick-c++-devel)")
     endif(MAGICKXXCONFIG)
     set(LIBRARIES ${LIBRARIES} ${ImageMagick_LIBRARIES})
-    include_directories(${ImageMagick_INCLUDE_DIRS})
+    include_directories(${ImageMagick_INCLUDE_DIRS} ${ImageMagick_MagickCore_INCLUDE_DIRS})
     set(MAGICK_LIBRARIES ${ImageMagick_LIBRARIES})
   else(ImageMagick_FOUND)
     message(FATAL_ERROR "ImageMagick is required but was not found.\n"
