### 2022-08-11 55415b0
was not what GraphicsMagick expected.
