### 2022-08-12 f516f85
change cache key to force recompilation of plplot with patch under linux
