+++ .github/workflows/main.yml
@@ -183,7 +183,7 @@ jobs:
         with:
           path: |
             C:\msys64\cache
-          key: cache-gdl-deps-windows-msys2-${{ matrix.arch }}-v8
+          key: cache-gdl-deps-windows-msys2-${{ matrix.arch }}-v9
       - name: Check Version of Latest GDL Workbench
         if: matrix.arch == 'x86_64'
         id: gdl_workbench
