### 2022-08-15 #4633 @mgalloy
@gkthiruvathukal I have completed my review and believe the paper is ready for publication.

I agree some direct examples of GDL usage and output in the README.md would be nice, but not required with all the other code examples in the unit tests, the .pro files in the library, and references to IDL examples.
