### 2022-08-18 e160b30
Tab widget was not (wxWidgets bug) compiling the best size of ALL tabs.
