### 2022-08-18 9e4a0b8
This version makes a correct installer on my win10.
