+++ CMakeLists.txt
@@ -48,7 +48,7 @@ include(TestBigEndian)
 
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_SOURCE_DIR}/CMakeModules)
 #### GDL is an interactive tool. Normally INTERACTIVE_GRAPHICS must be set
-set(INTERACTIVE_GRAPHICS ON CACHE BOOL "GDL: Disable check of existence of a suited graphics driver in plplot.)")
+set(INTERACTIVE_GRAPHICS ON CACHE BOOL "GDL: Disable checking existence of a suited graphics driver in plplot. Only file output may be available, and 3D graphics may NOT work.")
 
 #### Default cached values
 set(GDLDEV OFF CACHE BOOL "GDL: Enable GDL developer mode (includes work-in-progress tests)")
@@ -214,6 +214,7 @@ check_include_file(mpi.h HAVE_MPI_H)
 #     version of antlr in system-wide directories
 include_directories(src)
 
+#-------------------------------MANDATORY MODULES-------------------------------------
 if (NOT WIN32)
     # Ncurses MANDATORY for readline on POSIX
     # -DNCURSESDIR=DIR
@@ -284,6 +285,149 @@ else(ZLIB_FOUND)
         "(suitable Fedora/CentOS package: zlib-devel)")
 endif(ZLIB_FOUND)
 
+# GSL MANDATORY
+# -DGSLDIR=DIR
+set(CMAKE_PREFIX_PATH ${GSLDIR})
+find_package(GSL QUIET)
+set(HAVE_LIBGSL ${GSL_FOUND})
+set(HAVE_LIBGSLCBLAS ${GSL_FOUND})
+if(GSL_FOUND)
+    set(LIBRARIES ${LIBRARIES} ${GSL_LIBRARIES})
+    include_directories(${GSL_INCLUDE_DIR})
+else(GSL_FOUND)
+    message(FATAL_ERROR "Gnu Scientific library (1.7 or higher) and libgslcblas are mandatory.\n"
+    "Use -DGSLDIR=DIR to specify the gsl directory tree.\n"
+        "(suitable Debian/Ubuntu package: libgsl-dev)\n"
+        "(suitable Fedora/CentOS package: gsl-devel)")
+endif(GSL_FOUND)
+
+#RPC or XDR MANDATORY
+if(UNIX)
+    set(CMAKE_PREFIX_PATH ${RPCDIR})
+    find_package(RPC QUIET)
+    set(HAVE_RPC ${RPC_FOUND})
+    if(RPC_FOUND)
+        set(LIBRARIES ${LIBRARIES} ${RPC_LIBRARIES})
+        include_directories(${RPC_INCLUDE_DIR})
+    else(RPC_FOUND)
+        message(FATAL_ERROR "RPC support is mandatory.\n"
+        "Note that SunRPC has been removed in glibc-2.26 and later, "
+        "while being optional in earlier versions. Consider using the "
+        "recommended and more modern libtirpc instead.\n"
+        "Use -DRPCDIR=DIR to specify the rpc directory tree.\n")
+    endif(RPC_FOUND)
+elseif (WIN32)
+    set(CMAKE_PREFIX_PATH ${XDRDIR})
+    find_package(Xdr)
+    set(HAVE_LIBXDR ${XDR_FOUND})
+    if(XDR_FOUND)
+        set(LIBRARIES ${LIBRARIES} ${XDR_LIBRARIES})
+        include_directories(${XDR_INCLUDE_DIR})
+    else(XDR_FOUND)
[…]
