+++ scripts/build_gdl.sh
@@ -80,6 +80,7 @@ if [ ${BUILD_OS} == "Windows" ]; then
         readline zlib libpng gsl wxWidgets plplot libgd libtiff libgeotiff netcdf hdf4 hdf5 fftw proj msmpi udunits
         eigen3 eccodes glpk shapelib expat openssl
     )
+    #if you add something in MSYS2_PACKAGES_REBUILD you may have to add special lines in main.yml to push the product in /var/cache/pacman/pkg
     MSYS2_PACKAGES_REBUILD=(
         graphicsmagick plplot
     )
