+++ .github/workflows/main.yml
@@ -203,6 +203,8 @@ jobs:
           scripts/build_gdl.sh prep
           ls MINGW-packages/mingw-w64-graphicsmagick/
           mv MINGW-packages/mingw-w64-graphicsmagick/mingw-w64-*-graphicsmagick-*.tar.* /var/cache/pacman/pkg
+          ls MINGW-packages/mingw-w64-plplot/
+          mv MINGW-packages/mingw-w64-plplot/mingw-w64-*-plplot-*.tar.* /var/cache/pacman/pkg
           pacman -Q | tr ' ' '-' | sort > ~/snapshot2.txt
           echo "Setting up cache..."
           mkdir -p /cache
