### 2022-08-24 51a9649
Setting PLPLOT_DRV_DIR for Windows should enable access to our drivers. Removing copy of those drivers in bin for the windows case of build_gdl.sh
