### 2022-10-08 41bba5e
ensure drivers dlls are found both when compiling gdl under windows (using build_gdl.sh) AND when installed with the installer.
