+++ src/plotting.cpp
@@ -812,196 +812,198 @@ namespace lib
     //    large amount of data whitout duplicating all the arrays
     // trick 2/ when we have a NaN or and Inf, we realize the plot, then reset.
 
-    int GDL_POLYLINE_BUFFSIZE=500000; // idl default seems to be more than 2e6 !!
+    int GDL_POLYLINE_BUFSIZE=500000; // idl default seems to be more than 2e6 !!
 
-    if ( minEl<GDL_POLYLINE_BUFFSIZE ) GDL_POLYLINE_BUFFSIZE=append?minEl+1:minEl;
+    if ( minEl<GDL_POLYLINE_BUFSIZE ) GDL_POLYLINE_BUFSIZE=append?minEl+1:minEl;
     int i_buff=0;
-    PLFLT *x_buff=new PLFLT[GDL_POLYLINE_BUFFSIZE];
-    PLFLT *y_buff=new PLFLT[GDL_POLYLINE_BUFFSIZE];
+    PLFLT *x_buff=new PLFLT[GDL_POLYLINE_BUFSIZE];
+    PLFLT *y_buff=new PLFLT[GDL_POLYLINE_BUFSIZE];
 
     bool isBad=false;
+    if (append) //start with the old point
+    {
+      getLastPoint(a, x, y);
+      isBad = (!isfinite(x) || !isfinite(y) || isnan(x) || isnan(y));
+      if (!isBad) {
+        x_buff[i_buff]=x;
+        y_buff[i_buff]=y;
+        i_buff++;
+        append=true;
+      } else append=false;
+    } else {
+      if (!flag_x_const) x = static_cast<PLFLT> ((*xVal)[0]);
+      else x = x_ref;
+      if (!flag_y_const) y = static_cast<PLFLT> ((*yVal)[0]);
+      else y = y_ref;
+      x_buff[i_buff]=x;
+      y_buff[i_buff]=y;
+      i_buff++;
+    }
 
-    for ( SizeT i=0; i<minEl; ++i ) {
-      isBad=false;
-      if ( append ) //start with the old point
-      {
-        getLastPoint(a, x, y);
-        i--; //to get good counter afterwards
-        append=false; //and stop appending after!
-      }
-      else
-      {
-        if ( !flag_x_const ) x=static_cast<PLFLT>((*xVal)[i]);
-        else x=x_ref;
-        if ( !flag_y_const ) y=static_cast<PLFLT>((*yVal)[i]);
-        else y=y_ref;
-      }
-      
-      isBad=(!isfinite(x)|| !isfinite(y)||isnan(x)||isnan(y));
-      if ( isBad )
-      {
-        if ( i_buff>0 )
-        {
-          if ( line )
-          {
-            if (docolor) for (SizeT jj=0; jj< i_buff-1 ; ++jj)
-            {
-              a->Color ( ( *color )[plotIndex%color->N_Elements ( )], decomposed);
-              a->line(2, &(x_buff[jj]), &(y_buff[jj]));
-              plotIndex++;
-            }
-            else a->line(i_buff, x_buff, y_buff);
-          }
-          if (local_psym>0&&local_psym<8)
-          {
-            DLong oldStyl=gdlGetCurrentStyle();
+    // switches for fastness
+    bool hasNormalPsym=(local_psym > 0 && local_psym < 8);
+    bool hasCustomPsym=(local_psym==8);
+    bool hasHisto=(local_psym==10);
+    if (!docolor) { 
+      for (SizeT i = (append ? 0 : 1); i < minEl; ++i) {
[…]
