### 2022-10-25 #4633 @slayoo
@pjb7687 Apparently, we have already automation in place at Zenodo for archival of every release, so this is automatically done. The question is which version to reference in the paper? This can be "1.0.0", the last one, or perhaps a fresh one? Making a fresh one has the advantage that beforehand we can create the `.zenodo.json` file (like here: https://github.com/nipy/nipype/blob/master/.zenodo.json) which will instruct Zenodo how to populate the paper metadata. One problem is that CI fails as of today :( https://github.com/gnudatalanguage/gdl/actions/runs/3317371821
