### 2022-10-25 f1dbd13
this version correct some bugs related to how COLOR is supposed to be in PLOT,OPLOT and PLOTS. Refactoring should normally bring a significant speedup. A slowness due to colortable handling in plplot (usual bug) is rather abruptly avoided for wxWidgets backend by limiting the amount of drawn colors (should be invisible for many)
