### 2022-10-25 36b09c1
better logic for /continue and saved point value that should be NaN at start, not zero.
