### 2022-10-28 #4633 @pjb7687
@slayoo I think it makes more sense to me to deposit the version when we submitted the paper to JOSS, as this is the version of GDL that is actually described in the manuscript. We submitted it to JOSS on Jun 22, and I believe this one (https://github.com/gnudatalanguage/gdl/actions/runs/2613108929) is the closest working build. I would like to ask if you agree with it, if so I would deposit to Zenodo with manuscript.
