### 2022-10-28 #4633 @slayoo
@pjb7687 this commit is already published on Zenodo as a weekly release - we can just change its metadata: https://zenodo.org/record/6796032
