### 2022-10-30 #4633 @slayoo
@pjb7687 it's up to you, of course, but I wonder if having two Zenodo archives for the very same commit hash with differing "author lists" is not even more misleading?
