### 2022-11-03 #4633 @pjb7687
Hi all,

1. I made a tagged release 'v1.0.0-JOSS', based on commit 9416a4b998 (which is the last commit relevant to the JOSS manuscript). Here you can find the tag: https://github.com/gnudatalanguage/gdl/tree/v1.0.0-JOSS
2. The archive is published at Zenodo.
  - Link: https://zenodo.org/record/7275468#.Y2NJfnbP07c
3. I believe I have entered the metadata correctly but it would be nice if @slayoo would confirm it.
4. The DOI of the Zenodo release is: 10.5281/zenodo.7275468

Best,
Jeongbin
