### 2022-11-03 #4633 @slayoo
Thank you, @pjb7687!
I've checked the metadata against the `paper.md` (https://github.com/gnudatalanguage/gdl/blob/master/paper/paper.md) and clicked on all ORCID links - all seems to match.

@gkthiruvathukal, please do proceed, thanks!
