### 2022-11-07 40423c3
patch see #1377
getchar() returns an int, that should not be casted to unsigned chars (since EOF=-1) on machines where chars are unsigned.
