+++ src/plotting_windows.cpp
@@ -81,7 +81,9 @@ namespace lib {
 
     if (xSize < 0) xSize = 10000;
     if (ySize < 0) ySize = 10000;
-
+    //Limit size to max size permitted by wxWidgets : 32767 see #1399
+    xSize = (xSize > 32767)? 32767:xSize; 
+    ySize = (ySize > 32767)? 32767:ySize;
 
     //NOTE: xPos=-1 and yPos=-1 are when XPOS and YPOS options were not used!   
 
