### 2022-11-18 8de94e8
Passing an invalid clipbox will cause an absence of plot, but erasing the screen if MULTI=0 or just jumping to next plot if MULTI>1 . This is a better response to #1387
