+++ src/plotting_plots.cpp
@@ -186,7 +186,7 @@ namespace lib {
     bool prepareDrawArea(EnvT* e, GDLGStream* actStream) {
       //box defined in previous PLOT command, we pass in normalized coordinates w/clipping if needed 
       gdlSetSymsize(e, actStream); //set symsize BEFORE switching (TBC)
-      gdlSwitchToClippedNormalizedCoordinates(e, actStream, true); //inverted clip meaning
+      if (gdlSwitchToClippedNormalizedCoordinates(e, actStream, true)) return true; //inverted clip meaning
        return false;
     }
 
