### 2022-11-18 de048e2
try if removing explicit mention of python package in BREW_PACKAGES in build_gdl.sh helps with macOS CI problem
