### 2022-11-21 ce7f23b
insure X11 is included for OSX and UNIX as libx11 is available
