### 2022-11-21 fa2e634
insure compiles even with no X11 and no wxWidgets
