### 2022-11-21 e76254b
build our interactive graphic drivers only if the library is present
