### 2022-11-22 f484180
test if problem in comparing versions of numpy
