### 2022-11-22 317e275
remove python for macOSX, this numpy problem is not going to stop us.
