### 2022-11-22 f82b2fd
Bis repetita: remove python for macOSX, this numpy problem is not going to stop us.
