### 2022-11-22 6cdf427
solves a problem reported in https://github.com/gnudatalanguage/gdl/pull/1407#issuecomment-1322763693
