### 2022-11-22 5077781
ter repetita: remove python for macOSX, this numpy problem is not going to stop us.
