### 2022-11-24 d13c218
insure STOP,var behaviour same as IDL (i.e., silent)
