### 2022-11-29 cc2d730
enable .STEPOVER (.SO) .RETURN solves line number reported at  https://github.com/gnudatalanguage/gdl/issues/1403#issuecomment-1326490113
Possibly make loops faster in all cases especally in batch mode.
