+++ src/GDLInterpreter.cpp
@@ -73,40 +73,47 @@ GDLInterpreter::GDLInterpreter()
 	 RetCode retCode;
 	ProgNodeP statement_AST_in = (_t == ProgNodeP(ASTNULL)) ? ProgNodeP(antlr::nullAST) : _t;
 	
-	//    ProgNodeP& actPos = statement_AST_in;
+	
 	assert( _t != NULL);
 	ProgNodeP last;
 	_retTree = _t;
-	//  if( callStack.back()->GetLineNumber() == 0) 
-	//  if( _t->getLine() != 0) 
-	//      callStack.back()->SetLineNumber( _t->getLine());
 	
 	
 	try {      // for error handling
 		
+		//optimize speed: differentiate inner loops to avoid one externally implicit comparison on controlc for each loop (if possible)
+		if (interruptEnable) { //will test for sigControlC
+		{
 		do {
-		//                 if( _t->getLine() != 0) 
-		//                     callStack.back()->SetLineNumber( _t->getLine());
-		
 		last = _retTree;
-		
-		// track actual line number
-		callStack.back()->SetLineNumber( last->getLine());
-		
+		callStack.back()->SetLineNumber(last->getLine()); // track actual line number
 		retCode = last->Run(); // Run() sets _retTree
-		
-		}
-		while( 
-		_retTree != NULL && 
-		retCode == RC_OK && 
-		!(sigControlC && interruptEnable) && 
-		(debugMode == DEBUG_CLEAR));
-		
-		// commented out, because we are only at the last statement
-		// if( _retTree != NULL) 
-		//     last = _retTree;
-		
+		} while (_retTree != NULL && retCode == RC_OK && !(sigControlC) && (debugMode <= DEBUG_RETURN)); //loops if debug_clear or debug_return
+		if (_retTree != NULL) last = _retTree; //this is OK see https://github.com/gnudatalanguage/gdl/issues/1403#issuecomment-1326490113
 		goto afterStatement;
+		}
+		} else { //will not test for sigControlC 
+		{
+		do {
+		last = _retTree;
+		callStack.back()->SetLineNumber(last->getLine()); // track actual line number
+		retCode = last->Run(); // Run() sets _retTree
+		} while (_retTree != NULL && retCode == RC_OK && (debugMode <= DEBUG_RETURN)); //loops if debug_clear or debug_return
+		if (_retTree != NULL) last = _retTree; //this is OK see https://github.com/gnudatalanguage/gdl/issues/1403#issuecomment-1326490113
+		goto afterStatement;
+		}
+		}
+		// original single loop with all checks    
+		//    {
+		//      do {
+		//        last = _retTree;
+		//        callStack.back()->SetLineNumber(last->getLine()); // track actual line number
+		//        retCode = last->Run(); // Run() sets _retTree
+		//      } while (_retTree != NULL && retCode == RC_OK && !(sigControlC && interruptEnable) && (debugMode <= DEBUG_RETURN)); //loops if debug_clear or debug_return
+		//      if (_retTree != NULL) last = _retTree; //this is OK see https://github.com/gnudatalanguage/gdl/issues/1403#issuecomment-1326490113
+		//      goto afterStatement;
+		//    }
+		
 		
 		{
 		if (_t == ProgNodeP(antlr::nullAST) )
@@ -350,75 +357,59 @@ GDLInterpreter::GDLInterpreter()
 		}
 		}
 		
-		afterStatement:;
-		
+		afterStatement:
[…]
