### 2022-12-01 a88ac6e
insure all buttons in frame bar and hopefully no delay in showing the first window. Focus problems must be treated at window manager level, not in GDL
