### 2022-12-04 2fb2c2a
solve #1405 also probably plots appear faster.
