### 2022-12-06 885a15a
implement 'H5G_CREATE' command via 'h5g_create_fun()'.
