### 2022-12-07 ce28804
implement the 'H5G_GET_NMEMBERS' command using 'H5Giterate()' internally (as per IDL manual).
