+++ src/hdf5_fun.hpp
@@ -78,6 +78,7 @@ namespace lib {
   BaseGDL* h5a_read_fun( EnvT* e);
   BaseGDL* h5g_open_fun( EnvT* e);
   BaseGDL* h5g_create_fun( EnvT* e);
+  BaseGDL* h5g_get_nmembers_fun( EnvT* e);
   BaseGDL* h5g_get_objinfo_fun( EnvT* e);
   void h5a_write_pro( EnvT* e );
   void h5a_delete_pro( EnvT* e );
