### 2022-12-08 d8a1196
implement the 'H5G_GET_MEMBER_NAME' command using the 'H5Giterate()' function internally (as per IDL manual).
