### 2022-12-08 d64b22d
implement wrapper function for the 'H5G_GET_NUM_OBJS' command.
