### 2022-12-08 96be486
implement wrapper function for the 'H5G_GET_OBJ_NAME_BY_IDX' command.
