### 2022-12-08 1ade319
implement wrapper function for the 'H5G_SET_COMMENT' command.
