+++ src/hdf5_fun.hpp
@@ -84,6 +84,7 @@ namespace lib {
   void h5a_close_pro( EnvT* e );
   void h5t_close_pro( EnvT* e );
   void h5g_close_pro( EnvT* e );
+  void h5g_set_comment_pro( EnvT* e );
 
 } // namespace
 
