### 2022-12-08 1bfcaae
implement wrapper function for the 'H5G_GET_COMMENT' command.
