### 2022-12-09 bbb8945
in routine 'h5g_get_member_name_fun()', add guarding mechanism to free temporary string buffer when leaving the context.
