### 2022-12-09 129c6d5
in routine 'h5g_get_obj_name_by_idx_fun()', add guarding mechanism to free temporary string buffer when leaving the context.
