### 2022-12-09 f92876d
in routine 'h5g_get_comment_fun()', add guarding mechanism to free temporary string buffer when leaving the context.
