### 2022-12-09 2e8937f
add guarding mechanism to free temporary string buffers when leaving the context.
