### 2022-12-18 71be5fa
hm, the algorithm to get the same automatic ranges as IDL is bit more complicated than expected. Indeed, we need to insure that the !X or !Y.CRANGE, computed by FLOOR and CEIL on the (double precision) [X|Y]RANGE values do not 'jump' to the next interval due to numeric precision problems. I note that IDL behaves differently in this respect when the  [X|Y]RANGE is explicitely set to double precision: b=randomu(seed, 100000) & offset=2399.999d & plot, b, xrange=[offset, offset+24*24-1] will set lower bound to 2200 b=randomu(seed, 100000) & offset=2399.999 & plot, b, xrange=[offset, offset+24*24-
[…]
