### 2022-12-18 10f1093
test_usersym passes again: a oops in the draw_polyline() function was preventing to plot single PSYM points. So much for trying to speedup graphics! I note that plplot makes plots faster with a small buffersize, to be investigated in a future attepmpt to incorporate plplot to gdl.
