### 2022-12-19 995f335
implement the 'H5G_GET_NMEMBERS' and 'H5G_GET_MEMBER_NAME' commands, using 'H5Giterate()' internally (as per descriptioin in the IDL manual).
