### 2022-12-20 #4633 @danielskatz
@pjb7687 - I'm another track editor, helping out since the track editor here in out currently. I've proofread the paper, and have made a bunch of small suggestions in https://github.com/gnudatalanguage/gdl/pull/1443 - please merge this, or let me know what you disagree with, then we can proceed.
