### 2022-12-20 #4633 @danielskatz
@pjb7687 - once this command finishes, please check the new PDF to make sure everything still looks ok and let me know.  I'll do the same, then we can proceed to acceptance and publication
