### 2022-12-20 #4633 @danielskatz
@pjb7687 - the DOI isn't resolving for me yet, so I'm going to keep this open for now - I think this is related to [maintenance on Crossref](https://status.crossref.org) which should be complete and catch up in a few hours
