### 2022-12-20 #4633 @pjb7687
@danielskatz The DOI seems to work now, but the iframe shows me 404 instead of the pdf.
